feat: Add createdAt field to IMessageData interface and ChatBubbleProps
This commit adds the `createdAt` field to the `IMessageData` interface and `ChatBubbleProps` type. The `createdAt` field is used to store the timestamp when a message is created. This allows for displaying the creation time of each message in the chat bubble component.

The `createdAt` field is added to the `IMessageData` interface in the `types.ts` file, and to the `ChatBubbleProps` type in the `ChatBubble.types.ts` file.

This change is necessary to provide the necessary data for displaying the creation time of messages in the chat interface.